requirements.txt
Django
django-cors-headers
djangorestframework
pillow
python-dotenv
sqlparse
tzdata
psycopg2
gunicorn
requests


